+++ scripts/profile_memory.py
@@ -18,7 +18,7 @@ import numpy as np
 tracemalloc.start()
 
 
-def format_size(size_bytes: int) -> str:
+def format_size(size_bytes: float) -> str:
     """Format bytes as human-readable string."""
     for unit in ["B", "KB", "MB", "GB"]:
         if abs(size_bytes) < 1024.0:
